requirements.txt
requests
python-dotenv
apscheduler
streamlit
pandas
plotly
